Resset branch to 63d3ab5

+++ Uebungsprojekt/Controllers/AdministrationController.cs
@@ -24,18 +24,16 @@ namespace Uebungsprojekt.Controllers
         private readonly int max_allowed_filesize = (1024 * 1024) * 1; // Last multiplicator = mb
         private IMemoryCache cache;
         private HttpContext http_context;
-        private int user_id;
 
 
         /// <summary>
         /// Constructor for AdministrationController
         /// </summary>
         /// <param name="cache">IMemoryCache passed via Dependency Injection</param>
-        public AdministrationController(UserManager user_manager, IMemoryCache cache, IHttpContextAccessor http_context_accessor)
+        public AdministrationController(IMemoryCache cache, IHttpContextAccessor http_context_accessor)
         {
             this.cache = cache;
             this.http_context = http_context_accessor.HttpContext;
-            user_id = user_manager.GetUserIdByHttpContext(http_context_accessor.HttpContext);
         }
 
         /// <summary>
@@ -263,18 +261,13 @@ namespace Uebungsprojekt.Controllers
             return RedirectToAction("SimulationInfrastructure");
         }
 
-        [HttpGet, ActionName("DeleteSimulationLocation")]
+        [HttpGet]
         public ActionResult DeleteSimulationLocation(int id)
         {
             LocationDaoImpl locationDao = new LocationDaoImpl(cache);
             SimulationInfrastructure infrastructure = GetSimulationInfrastructureFromCookie();
-            locationDao.Delete(id, infrastructure.location_dao_id);
-            return RedirectToAction("SimulationInfrastructure");
-        }
-
-        [HttpPost, ActionName("DeleteSimulationLocation")]
-        public ActionResult DeleteSimulationLocationConfirmed(int id)
-        {
+            if (infrastructure != null)
+                locationDao.Delete(id, infrastructure.location_dao_id);
             return RedirectToAction("SimulationInfrastructure");
         }
 
@@ -299,7 +292,6 @@ namespace Uebungsprojekt.Controllers
             SimulationInfrastructure infrastructure = GetSimulationInfrastructureFromCookie();
             if (infrastructure == null)
                 return RedirectToAction("SimulationInfrastructure");
-            
             LocationDao location_dao = new LocationDaoImpl(cache);
             ChargingZoneDaoImpl charging_zone_dao = new ChargingZoneDaoImpl(cache);
             
@@ -310,24 +302,19 @@ namespace Uebungsprojekt.Controllers
                 infrastructure.charging_zone_dao_id
                 );
 
-            Console.Out.WriteLine(charge.name);
             return RedirectToAction("SimulationInfrastructure");
         }
 
-        [HttpGet, ActionName("DeleteSimulationChargingZone")]
+        [HttpGet]
         public ActionResult DeleteSimulationChargingZone(int id)
         {
             ChargingZoneDaoImpl chargingZoneDao = new ChargingZoneDaoImpl(cache);
             SimulationInfrastructure infrastructure = GetSimulationInfrastructureFromCookie();
-            chargingZoneDao.Delete(id, infrastructure.charging_zone_dao_id);
+            if (infrastructure != null)
+                chargingZoneDao.Delete(id, infrastructure.charging_zone_dao_id);
             return RedirectToAction("SimulationInfrastructure");
         }
-
-        [HttpPost, ActionName("DeleteSimulationChargingZone")]
-        public ActionResult DeleteSimulationChargingZoneConfirmed(int id)
-        {
-            return RedirectToAction("SimulationInfrastructure");
-        }
+        
 
         [HttpGet]
         public IActionResult AddSimulationChargingColumn()
@@ -365,18 +352,15 @@ namespace Uebungsprojekt.Controllers
             return RedirectToAction("SimulationInfrastructure");
         }
 
-        [HttpGet, ActionName("DeleteSimulationChargingColumn")]
+        [HttpGet]
         public ActionResult DeleteSimulationChargingColumn(int id)
         {
-            ChargingColumnDaoImpl chargingColumnTypeDao = new ChargingColumnDaoImpl(cache);
             SimulationInfrastructure infrastructure = GetSimulationInfrastructureFromCookie();
+            if (infrastructure == null)
+                return RedirectToAction("SimulationInfrastructure");
+            
+            ChargingColumnDaoImpl chargingColumnTypeDao = new ChargingColumnDaoImpl(cache);
             chargingColumnTypeDao.Delete(id, infrastructure.charging_column_dao_id);
-            return RedirectToAction("SimulationInfrastructure");
-        }
-
-        [HttpPost, ActionName("DeleteSimulationChargingColumn")]
-        public ActionResult DeleteSimulationChargingColumnConfirmed(int id)
-        {
             return RedirectToAction("SimulationInfrastructure");
         }
 
@@ -385,6 +369,8 @@ namespace Uebungsprojekt.Controllers
         {
             VehicleDaoImpl vehicle_dao = new VehicleDaoImpl(cache);
             SimulationConfig config = GetSimulationConfigFromCookie();
+            if (config == null)
+                return RedirectToAction("SimulationConfig");
             AddSimulationVehicleViewModel svvm = new AddSimulationVehicleViewModel(vehicle_dao.GetAll(), config);
             return View(svvm);
         }
@@ -393,6 +379,9 @@ namespace Uebungsprojekt.Controllers
         public IActionResult AddSimulationVehicle(int vehicle_id, int count)
         {
             SimulationConfig config = GetSimulationConfigFromCookie();
+            if (config == null)
+                return RedirectToAction("SimulationConfig");
+            
             VehicleDaoImpl vehicle_dao = new VehicleDaoImpl(cache);
             for(int i = 0; i < count; i++)
             {
@@ -401,38 +390,59 @@ namespace Uebungsprojekt.Controllers
             return RedirectToAction("AddSimulationVehicle");
         }
 
-        [HttpGet, ActionName("DeleteSimulationVehicle")]
+        [HttpGet]
         public ActionResult DeleteSimulationVehicle(int id)
         {
             SimulationConfig config = GetSimulationConfigFromCookie();
-            VehicleDaoImpl vehicle_dao = new VehicleDaoImpl(cache);
-            config.vehicles.Remove(vehicle_dao.GetById(id));
+            if (config != null)
+            {
+                VehicleDaoImpl vehicle_dao = new VehicleDaoImpl(cache);
+                config.vehicles.Remove(vehicle_dao.GetById(id));
+            }
             return RedirectToAction("AddSimulationVehicle");
         }
 
-        [HttpPost, ActionName("DeleteSimulationVehicle")]
-        public ActionResult DeleteSimulationVehicleConfirmed(int id)
-        {
-            return RedirectToAction("AddSimulationVehicle");
-        }
-
-        [HttpGet, ActionName("DeleteAllSimulationVehicles")]
+        [HttpGet]
         public ActionResult DeleteAllSimulationVehicles()
         {
             SimulationConfig config = GetSimulationConfigFromCookie();
+            if (config == null)
+                return RedirectToAction("SimulationConfig");
+            
             config.vehicles.Clear();
             return RedirectToAction("AddSimulationVehicle");
         }
 
-        [HttpPost, ActionName("DeleteAllSimulationVehicles")]
-        public ActionResult DeleteAllSimulationVehiclesConfirmed()
-        {
-            return RedirectToAction("AddSimulationVehicle");
-        }
-
+        [HttpGet]
         public IActionResult AddRushHours()
         {
-            return View();
+            SimulationConfig config = GetSimulationConfigFromCookie();
+            if (config == null)
+                return RedirectToAction("SimulationConfig");
+            
+            return View(config);
+        }
+        
+        [HttpPost]
+        public IActionResult AddRushHours(DayOfWeek weekday, TimeSpan timespan)
+        {
+            SimulationConfig config = GetSimulationConfigFromCookie();
+            if (config == null)
+                return RedirectToAction("SimulationConfig");
+            
+            config.rush_hours.Add(new Tuple<DayOfWeek, TimeSpan>(weekday, timespan));
+            return View(config);
+        }
+        
+        [HttpGet]
+        public IActionResult DeleteRushHours(int index)
+        {
+            SimulationConfig config = GetSimulationConfigFromCookie();
+            if (config == null)
+                return RedirectToAction("SimulationConfig");
+            
+            config.rush_hours.RemoveAt(index);
+            return RedirectToAction("AddRushHours");
         }
 
         /// <summary>
@@ -449,7 +459,6 @@ namespace Uebungsprojekt.Controllers
         [HttpPost]
         public IActionResult CreateChargingZone(ChargingZone charge, int location_id)
         {
-            Console.WriteLine(charge.name+ " performance: "+charge.overall_performance);
             ChargingZoneDaoImpl charging_zone = new ChargingZoneDaoImpl(cache);
             LocationDaoImpl location_dao = new LocationDaoImpl(cache);
             charging_zone.Create(charge.name,charge.overall_performance, location_dao.GetById(location_id, 0), 0);
@@ -486,7 +495,6 @@ namespace Uebungsprojekt.Controllers
         public IActionResult AddChargingColumn(int charging_zone_id, int charging_column_type_id )
         {
             //TODO: ERRORHANDLING!
-            Console.WriteLine(charging_column_type_id);
             ChargingColumnTypeDaoImpl charging_column_type = new ChargingColumnTypeDaoImpl(cache);
             ChargingColumnDaoImpl charging_columnd_dao = new ChargingColumnDaoImpl(cache);
             ChargingZoneDaoImpl charging_zone_dao = new ChargingZoneDaoImpl(cache);
@@ -650,8 +658,7 @@ namespace Uebungsprojekt.Controllers
             if (ModelState.IsValid && vehicle.connector_types != null)
             {
                 VehicleDao vehicle_dao = new VehicleDaoImpl(cache);
-                UserDao user_dao = new UserDaoImpl(cache);
-                vehicle_dao.Create(vehicle.model_name, vehicle.capacity, vehicle.connector_types, user_dao.GetById(user_id)); // TODO: soll hier vehicle mit admin user verbunden werden?
+                vehicle_dao.Create(vehicle.model_name, vehicle.capacity, vehicle.connector_types);
                 return RedirectToAction("Vehicles");
             }
             return RedirectToAction("Vehicles");
